fix: Disambiguate conflicting test modules
